Title: 사칙연산 함수, Time: 41ms, Memory: 16MB, Status: Passed - CodeTree

+++ 250112/사칙연산 함수/quadratic-operations-function.py
@@ -1,16 +1,16 @@
 def plus(a, c):
-    return a + c
+    return int(a + c)
 
 def minus(a, c):
-    return a - c
+    return int(a - c)
 
 def mul(a, c):
-    return a * c
+    return int(a * c)
 
 def div(a, c):
     if c == 0:
         return "0으로 나눌 수 없습니다."
-    return a / c
+    return int(a / c)
 
 def main():
     a, o, c = input().split()
